+++ lib/blazer/engine.rb
@@ -3,22 +3,24 @@ module Blazer
     isolate_namespace Blazer
 
     initializer "blazer" do |app|
-      if defined?(Sprockets) && Sprockets::VERSION.to_i >= 4
-        app.config.assets.precompile += [
-          "blazer/application.js",
-          "blazer/application.css",
-          "blazer/glyphicons-halflings-regular.eot",
-          "blazer/glyphicons-halflings-regular.svg",
-          "blazer/glyphicons-halflings-regular.ttf",
-          "blazer/glyphicons-halflings-regular.woff",
-          "blazer/glyphicons-halflings-regular.woff2",
-          "blazer/favicon.png"
-        ]
-      else
-        # use a proc instead of a string
-        app.config.assets.precompile << proc { |path| path =~ /\Ablazer\/application\.(js|css)\z/ }
-        app.config.assets.precompile << proc { |path| path =~ /\Ablazer\/.+\.(eot|svg|ttf|woff|woff2)\z/ }
-        app.config.assets.precompile << proc { |path| path == "blazer/favicon.png" }
+      if app.config.respond_to?(:assets) && defined?(Sprockets)
+        if Sprockets::VERSION.to_i >= 4
+          app.config.assets.precompile += [
+            "blazer/application.js",
+            "blazer/application.css",
+            "blazer/glyphicons-halflings-regular.eot",
+            "blazer/glyphicons-halflings-regular.svg",
+            "blazer/glyphicons-halflings-regular.ttf",
+            "blazer/glyphicons-halflings-regular.woff",
+            "blazer/glyphicons-halflings-regular.woff2",
+            "blazer/favicon.png"
+          ]
+        else
+          # use a proc instead of a string
+          app.config.assets.precompile << proc { |path| path =~ /\Ablazer\/application\.(js|css)\z/ }
+          app.config.assets.precompile << proc { |path| path =~ /\Ablazer\/.+\.(eot|svg|ttf|woff|woff2)\z/ }
+          app.config.assets.precompile << proc { |path| path == "blazer/favicon.png" }
+        end
       end
 
       Blazer.time_zone ||= Blazer.settings["time_zone"] || Time.zone