+++ lib/blazer/statement.rb
@@ -33,9 +33,12 @@ module Blazer
           end
 
           unless value.is_a?(ActiveSupport::TimeWithZone)
-            if value =~ /\A\d+\z/
-              value = value.to_i
-            elsif value =~ /\A\d+\.\d+\z/
+            if value.match?(/\A\d+\z/)
+              # check no leading zeros (when not zero)
+              if value == value.to_i.to_s
+                value = value.to_i
+              end
+            elsif value.match?(/\A\d+\.\d+\z/)
               value = value.to_f
             end
           end