+++ test/adapters/mysql_test.rb
@@ -4,7 +4,7 @@ class MysqlTest < ActionDispatch::IntegrationTest
   include AdapterTest
 
   def data_source
-    "mysql"
+    ENV["MYSQL_ADAPTER"] || "mysql2"
   end
 
   def test_run
@@ -90,6 +90,16 @@ class MysqlTest < ActionDispatch::IntegrationTest
   def test_binary
     # checks for successful response
     run_statement "SELECT UNHEX('F6'), 1", format: "html"
+  end
+
+  def test_binary_output
+    assert_raises(CSV::InvalidEncodingError) do
+      assert_result [{"hello" => "0xF6"}], "SELECT UNHEX('F6') AS hello"
+    end
+  end
+
+  def test_json_output
+    assert_result [{"json" => '{"hello": "world"}'}], %!SELECT JSON_OBJECT('hello', 'world') AS json!
   end
 
   private