+++ test/forecasting_test.rb
@@ -1,6 +1,10 @@
 require_relative "test_helper"
 
 class ForecastingTest < ActionDispatch::IntegrationTest
+  def setup
+    Blazer::Query.delete_all
+  end
+
   def test_prophet
     skip unless ENV["TEST_PROPHET"]
 
@@ -14,6 +18,8 @@ class ForecastingTest < ActionDispatch::IntegrationTest
   end
 
   def assert_forecast(forecasting)
+    skip unless postgresql?
+
     Blazer.stub(:forecasting, forecasting) do
       query = create_query(statement: "SELECT current_date + n AS day, n FROM generate_series(1, 30) n")
       run_query query.statement, query_id: query.id, forecast: "t"